+++ src/main/java/cn/mauth/account/dao/VoucherLineDao.java
@@ -5,5 +5,5 @@ import cn.mauth.account.common.domain.settings.VoucherLine;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VoucherLineDao extends BaseDao<VoucherLine,Integer>{
+public interface VoucherLineDao extends BaseDao<VoucherLine,Long>{
 }
